With the new cut-down index, my previous upper-bound estimate on index file size is now laughably inaccurate, and worse still it may now prove to be the limiting factor on the size of index that can be handled (if it causes address space to run out long before actual memory is in danger). I've tried and failed to think of a way of estimating the upper bound more accurately in the new circumstances, so instead I'm giving up: instead of padding out the index file to the maximum size that could possibly be needed, we extend it bit by bit, re-mmapping as we go.
I've also retired --full, because I now realise it's utterly
unnecessary: the only thing it might have been useful for would have
been running queries based on a single file - and if we wanted to do
that, we could simply pick one entry out of the trie and not go to
the index at all. So full indexes are a thing of the past, and good
riddance.

[originally from svn r8288]