Escape dots at the start of pathname components. (Mostly for HTML multifile output mode, but I think it's at worst harmless to leave it enabled in web server contexts too.)
Also I've just realised that if you type in a URL and guess wrongly
which characters agedu will have quoted, the parse function will work
it out anyway and give you a redirect to the canonical version of the
path. This was more or less accidental - an unanticipated consequence
of my parse-reformat-redirect strategy - but it seems thoroughly
useful. Add a comment pretending I meant to do it all along :-)

[originally from svn r9153]