Rearrange documentation of -S, -L and -D.
I previously documented -D and -L first, as obvious inverses of one
another (turn an index file into a textual dump and vice versa), and
then mentioned -S as an afterthought (go straight from scan to dump
without the intermediate index file). But a conversation with a user
suggested to me that this is actually a silly way to document this
cluster of options, because the most likely combination of them to be
actually useful is not -D and -L, but -S and -L - if you want to scan
one machine and end up with an index on another, those options are the
pair that get you most efficiently to that state. -D is actually for
the more specialist situation in which you've already built an index
on one machine but then _change your mind_ about where it should live.

The initial intro section now contains an example of -S and -L in
action, and the per-option documentation has been switched round to
mention those two first with simple descriptions and relegate -D to
the position of 'afterthought with strange extra caveats' previously
occupied by -S.