Add a --numeric option, modifying HTML multifile output.
This causes the HTML files (other than index.html) to be named after
the numeric node ids in the index file of the corresponding pathname.

This is exactly the thing I moved away from in commit f2caa217f, with
a comment that it could be easily enough re-enabled by an option if
anyone needed it. Well, this week a user pointed out that in the new
default mode of naming HTML files after the entire pathname they
index, it's quite possible for a deep directory tree on input to cause
agedu -H to exceed the filesystem's name-length limit on output. And
despite the other downsides of the numeric naming system, it does have
the advantage of being immune to _that_ problem - so here's an option
to revert to it.