Revise parsing of /proc/net/tcp[6] for magic HTTP auth.
I had assumed that the address and uid fields of each line of those
files would always occur at fixed column numbers. But in fact, that's
not true, because checking the kernel source, the initial sequence
number on the line is printed using a "%4d" printf format, which means
that if there are enough simultaneous connections on a given machine
then it could overflow to 5 digits and cause all the subsequent column
positions to be off by one.

Now we parse on a word-by-word basis, which should be more robust.