Give a sensible error message about misordered dumps.
Since creataing dump files is not agedu's sole privilege - users can
construct their own, and may get it wrong - it's not polite to respond
with an assertion failure to a dump file in the wrong sorting order.
Now we give a proper error message citing the ilne of the file where
the problem occurred.