Tolerate duplicate file names returned from readdir().
A user reports that this can happen using the 'gluster' network
filesystem, and also I think it's at least _imaginable_ that it might
happen if a directory is updated by another process while agedu is
half way through readdir()ing it - at least, POSIX doesn't seem to
contain any text promising that that won't happen, so it's possible in
principle even if most implementations in practice manage to avoid it.

If it does happen, we just have to treat it the same as there being
one copy of that file name in the directory (since any attempt to open
the file or even stat it for its inode would have to use the name as
the lookup key).