Add an option to use st_size instead of st_blocks.
A user reported that at least one NAS device returns obvious lies for
st_blocks (and not in any way that can be easily corrected for, like
multiplying them by a different factor). So I add this option as a
last-ditch workaround, though I don't see any reason to recommend it
if there's any alternative.