Rename the [files] entry in HTML show-files mode.
In --files mode, each individual file in the directory you're looking
at is shown as its own line in the HTML report. But a line labelled
[files] still shows up as well. Why? Because that line is the only
part of the overall size of the directory being listed that was not
moved into a separate line - namely, the directory inode itself.

I don't think that's actually _wrong_ behaviour. It's just listed with
a confusing and silly name. Rename it to [directory] in that situation.