Fix assertion failure in 'agedu --help'.
The help output code enforces by assertion that if you add an option
with a value, you supply a metavariable name to write in the help
after the option name. But I forgot to add one for the new --launch
option, and didn't try out the --help option before pushing, so of
course the assertion didn't catch it for me.