Fix assertion failure in index_order_stat.
I ran into a case today where 'size = node->totalsize * f;' can make
size bigger than node->totalsize started out. What happened was that f
was equal to 1.0, so the effect was just that of casting the 64-bit
integer node->totalsize to and from a double; and the integer value
was large enough that it couldn't be represented _exactly_ as a
double, so it was rounded - and, by chance, rounded _up_, failing the
assertion.

It would be nice to fix this by doing a more rigorous evaluation, but
honestly, I can't be bothered; I think these order statistics are too
approximate to worry about. So I just clip the value back into range.

(In any case, this can only happen with a filesystem size exceeding
2^53 bytes = 8PB. Or rather, with a scan file _claiming_ that's what
you've got - which is what happened in my case, due to the scan file
corruption caused by the bug fixed in the previous commit.)