Convert build system to cmake.
I've been doing this to the rest of my software projects recently,
mostly because CMake is friendlier than autotools to projects that
build for Windows, because it can treat Windows as a first-class
citizen. In this case, agedu _itself_ is a Unix-only program, but it
comes with the auxiliary ageduscan.exe that builds for Windows - so
now I can write a CMakeLists that builds the Unix program on Unix, and
builds the Windows program on Windows.

Changes in the actual source code are minimal:

 - replace #ifdef HAVE_FOO with #if HAVE_FOO, since cmake's
   configuration header likes to define things unconditionally and set
   them to 0 or 1

 - change the versioning system, because autotools isn't defining
   PACKAGE_VERSION any more

 - stop using SELECT_TYPE_ARG1. SELECT_TYPE_ARG234 etc, because only
   autoconf defined those. (With any luck they're not needed anyway in
   the modern world, since the types of select arguments are defined
   by POSIX; I expect those were one of autoconf's especially
   historical provisions.)