Fix infinite loop in round_and_format_age.
This occurs if an atime in agedu's index has value zero, or close to
it. I found it while manually constructing an agedu dump file from
other data: I accidentally set _all_ the atime fields to zero.

round_and_format_age works by making 'struct tm' more and more years
in the past until it finds one earlier than the given input value, but
stores the results in an unsigned long long (that being the format of
atimes in the dump file). When the test 'struct tm' had the year 1970,
it was still newer than time zero; when it had 1969, mktime returned a
negative time_t, which turned into a huge one, which _still_ compared
greater than the target time.

Now we spot the case where unsigned integer underflow has occurred,
and use it as another reason to terminate the loop.